Add files via upload
Program: Manager of Code

Description:
This program was created for practice and learning purposes. It includes basic functionalities such as reading data, processing it, and outputting results to a file.

The application can be modified and extended according to the user’s needs.

Programming Language:
- Python 3

Creation Date: September 2025

License:
Open source – you are free to use, modify, and distribute the code.